requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
telebot~=0.0.5
pytest~=8.2.2
pip~=24.0
attrs~=23.2.0
wheel~=0.43.0
setuptools~=70.0.0
packaging~=24.1
selenium~=4.22.0
selenium~=4.22.0